// @/components/ui/command/types.ts
type NavLinkProps = AnchorHTMLAttributes<HTMLAnchorElement>;

export type CategoryCommand = {
  /**
   * Icon to be displayed next to the command text
   */
  icon?: ReactElement;
  /**
   * The text displayed on the command
   */
  text: string;
  /**
   * The action to be performed
   */
  perform?: () => void;
  /**
   * The URL to be opened
   */
  href?: string;
  /**
   * Whether or not that URL should open in a new tab
   */
  newTab?: boolean;
  /**
   * Keywords for the command that will make it show up when the user searches something
   */
  keywords?: string;
  /**
   * Keyboard shortcuts which can trigger the command OUTSIDE the command menu
   *
   */
  shortcuts?: Shortcut;
  /**
   * Whether or not to close this menu when the function is called
   */
  closeOnComplete?: boolean;
  /**
   * Allow for custom HTML to be passed as the anchor property
   */
  anchor?: FC<NavLinkProps>;
};

export type Shortcut = {
  /**
   * Key that will be used in conjunction with the shortcut
   */
  modifier?: "shift" | "ctrl" | "alt" | "meta";
  /**
   * The key(s) that will trigger the function
   */
  keys: [string, string?];
};